22f1001871/BACKEND/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.2.0
billiard==4.2.0
bleach==6.1.0
blinker==1.8.2
Brotli==1.1.0
cachelib==0.9.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Bleach==0.0.2
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.53.1
greenlet==3.0.3
html5lib==1.1
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
kombu==5.3.7
MarkupSafe==2.1.5
matplotlib==3.9.1
numpy==2.0.1
packaging==24.1
passlib==1.7.4
pillow==10.4.0
prompt_toolkit==3.0.47
pycparser==2.22
pydyf==0.11.0
pyparsing==3.1.2
pyphen==0.15.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.7
requests==2.32.3
setuptools==71.1.0
six==1.16.0
SQLAlchemy==2.0.31
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.3
wheel==0.43.0
WTForms==3.1.2
zopfli==0.2.3
